fix(PET-1098): Fix for intinite connection retries
When used in celery kombu fails to reconnect to rabbitmq after it loses connection enough times when task are being processed. Eventually the master process would attempt to connect to rabbitmq and then close the connection after a very brief moment and do that until an error occurred, which would break the infinite loop.
This is cloning changes from this PR which is not yet merged: https://github.com/celery/kombu/pull/2275
I have verified that the issue of infinite connection creation is fixed, but other celery problems are still there. It is still an improvement.